Switched local port from 6502 to 6503.
Opening a W5100 UDP socket on some port means that the W5100 MAC raw socket doesn't receive incoming packets on that port anymore - the W5100 obviously doesn't deliver incoming packets to multiple sockets. However one needs to internalize that the "high level" socket actually acts like a filter for the MAC raw socket.

Uther II Drive is supposed to be cooperating with the ADTPro client, which has a W5100 MAC raw socket listen on port 6502 to replies from the ADTPro server. So Uther II Drive must have its W5100 UDP socket listen on a different port.

Note: Currently the ADTPro server presumes that all requests come from the same port. This needs to be changed in order to allow the ADTPro server to reply both to the ADTPro client (requesting on port 6502) and Uther II Drive (requesting on port 6503).